# curso/docs/swagger.yaml
basePath: /
definitions:
  api.Error:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8082
info:
  contact:
    email: atendimento@fdqf.com.br
    name: Gustavo P Gialluisi
    url: http://www.fdqf.com
  description: Microserviço modelo para cadastro de cursos
  license:
    name: FDQF License
    url: http://license.fdqf.com.br
  termsOfService: http://swagger.io/terms/
  title: FDQF - Microserviço de Cursos
  version: 0.0.1
paths:
...
  /users:
    post:
      consumes:
      - application/json
      description: Create user
      parameters:
      - description: user request
        in: body
        name: request
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Create user
      tags:
      - users
  /users/generate_token:
    post:
      consumes:
      - application/json
      description: Get a user JWT
      parameters:
      - description: user credentials
        in: body
        name: request
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a user JWT
      tags:
      - users
swagger: "2.0"